Clean up of base version
