+++ README.md
@@ -1,3 +1,49 @@
-# SoftwareProjects_AnimalRecognition
+# Animal Recognition Prototype
 Software Projects - Elanco Group 3
 
+Version: 0.1
+February 2023
+Professional Software Projects
+Liam, James, Sam, Mohammad
+
+
+------------------
+1.  INTRODUCTION
+------------------
+This prototype is a web application for our client Elanco. It takes a use input image and uses the Google Cloud Vision API to scan the image and return labels. We then manage the data for a back end service, and display refactored data to the web user.
+
+------------------
+2.  DOCUMENTATION
+------------------
+This application is made by Liam, James, Sam, Mohammad, and should only be used and/or distributed by those authorised by themselves, Elanco, or Sheffield Hallam University.
+
+------------------
+3. DEPENDENCIES
+------------------
+This web application has the following dependencies:
+
+(a) Google cloud Vision API: ^3.1.0
+    https://cloud.google.com/vision/docs/libraries
+    
+(b) Embedded JavaScript Template: "^3.1.8"
+    https://ejs.co
+
+(c) Node.js: ^18.13.0
+    https://nodejs.org/en/
+
+(d) Express.js: ^4.18.2
+    https://expressjs.com
+
+(e) Universally Unique Identifiers: ^9.0.0
+    https://www.npmjs.com/package/uuid
+  
+------------------
+4.  RUNNNING
+------------------
+
+
+
+------------------
+5.  SUPPORT
+------------------
+Following and further issues, please go through this readme file again, if the program still does not function get in touch with any of the authors directly.
