+++ client/views/index.ejs
@@ -1,0 +1,43 @@
+<!DOCTYPE html>
+<html lan="en" and dir="Itr">
+<head>
+    <meta charset="utf-8">
+    <title> Interactive Login Form</title>
+    <link rel="stylesheet" href="index.css">
+    <script src ="admin-login.js"> </script>
+
+
+
+</head>
+
+<body>
+    <header>
+        <div class="container">
+          <div class="logo">
+            <img src="frontEndImages/elanco.png" alt="Logo" />
+          </div>
+
+          <div class="links">
+            <ul>
+              <li><a href="#" class="btn">Sign up</a></li>
+            </ul>
+          </div>
+
+          <div class="overlay"></div>
+        </header>
+        <div class="box">
+          <h2>image dropbox</h2>
+          <div class="rectangle">
+            
+            <h3 style="position: absolute;top: 30px;left: 55px;">Press to take/upload photo</h3>
+          </div>
+          <div class="links">
+            <ul>
+              <li><a href="results.html" class="btn2">Search</a></li>
+            </ul>
+          </div>
+         </div>
+        </div>
+       
+      </body>
+      </html>
