+++ package.json
@@ -1,13 +1,13 @@
 {
-  "name": "expresstest",
-  "version": "1.0.0",
+  "name": "SP_AnimalRecognition",
+  "version": "0.0.1",
   "description": "",
-  "main": "index.js",
+  "main": "main.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
-  "author": "",
+  "author": "Liam Hammond",
   "license": "ISC",
   "dependencies": {
     "@google-cloud/vision": "^3.1.0",
