--- public/results.css
@@ -1,94 +1,0 @@
-a {
-    text-decoration: none;
-  }
-  header {
-    position: relative;
-    z-index: 70;
-  }
-  
-  header .container {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
-  
-  .overlay {
-    display: none;
-  }
-  
-  .logo {
-    display: flex;
-    align-items: center;
-    cursor: pointer;
-  }
-  
-  .logo img {
-    width: 150px;
-    height: 75px;
-    margin-right: 0.6rem;
-    margin-top: -0.6rem;
-    position: relative; left: 25px;
-  }
-  .links ul {
-    display: flex;
-    list-style: none;
-    align-items: center;
-  }
-  
-  .links a {
-    color: white;
-    margin-left: 4.5rem;
-    display: inline-block;
-    transition: 0.3s;
-    
-  }
-  
-  .btn{
-    position: relative; right: 30px;
-    border: 0;
-    background-color: #0367fd;
-    display: block;
-    margin: 20px auto;
-    text-align: center;
-    padding:14px 35px;
-    outline: none;
-    color: white;
-    border-radius: 24px;
-    transition:0.25px;
-    cursor:pointer;
-    
-  }
-  .box{
-    width:750px;
-    height: 375px;
-    padding:30px;
-    position: absolute;
-    top:70%;
-    left:50%;
-    transform: translate(-50%,-50%);
-    background:#C8C8C8;
-    text-align:center;
-    border-radius: 40px;
-    opacity: 100%;
-
-  }
-  .box h4
-{
-  color: black;
-  text-transform: uppercase;
-  font-weight: bold;
-  position: fixed;
-  font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
-  font-size: 25px;
-  top: 10px;
-  right: 350px;
-}
-.rectangle{
-    width:625px;
-    height:300px;
-    position: relative;
-    left: 425px;
-    background:white;
-    top: 375px;
-    border-radius: 24px;
-}
