--- views/admin-login.ejs
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html lan="en" and dir="Itr">
-<head>
-    <meta charset="utf-8">
-    <title> Interactive Login Form</title>
-    <link rel="stylesheet" href="style.css">
-    <script src ="admin-login.js"> </script>
-
-
-
-</head>
-
-<body>
-    <header>
-        <div class="container">
-          <div class="logo">
-            <img src="images/elanco.png" alt="Logo" />
-          </div>
-
-          <div class="links">
-            <ul>
-              <li><a href="#" class="btn">Sign up</a></li>
-            </ul>
-          </div>
-
-          <div class="overlay"></div>
-        </header>
-<form class="box" action="admin-login.html" method="POST">
-
-<h1>
-    login
-</h1>
-<input type="text" name="" placeholder="Enter Username" id="username">
-<input type="password" name="" placeholder="Enter Password" id="password">
-<input type="submit" name="" value="Login"  onclick="validate()">
-
-
-</form>
-
-
-</body>
-
-
-
-</html>
