--- views/home.ejs
@@ -1,27 +1,0 @@
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <link rel="stylesheet" href="/style.css">        
-</head>
-<body>
-    Welcome to the homepage
-    <form action="/result" method="POST">
-        <input type="text" name="colour">  
-        <button>Submit</button> 
-    </form>
-
-    <% if (isRaining) %>
-        <p>It is raining</p>
-
-    <%= isRaining %>
-
-    <p>Current pets:</p>
-    <% pets.forEach(pet => { %>
-        <li>Name: <%= pet.name %> is a <%= pet.animal %></li>
-    <% }) %>
-
-</body>
-</html>
